Refactor /crypto
