+++ crypto/Cargo.toml
@@ -14,6 +14,7 @@ license-file = "../mit-license.md"
 [features]
 default = [ ]
 bench = [ "criterion" ]
+arkworks = [ "dep:ruint" ]
 
 [[bench]]
 name = "criterion"
@@ -29,7 +30,7 @@ ark-ff = { version = "0.3.0", features = ["parallel", "asm"] }
 hex = "0.4.3"
 rand = "0.8.5"
 rayon = "1.5.3"
-ruint = { version = "1.3.0", features = ["ark-ff"] }
+ruint = { version = "1.3.0", features = ["ark-ff"], optional = true }
 serde = { version = "1.0.144", features = ["derive"] }
 serde_json = "1.0.85"
 thiserror = "1.0.34"
