+++ crypto/src/powers.rs
