+++ crypto/src/transcript.rs
